minor style updates

+++ cgogn/core/cmap/cmap3.h
@@ -37,9 +37,9 @@ class CMap3_T : public CMap2_T<MAP_TRAITS, MAP_TYPE>
 {
 public:
 
-	static const int32 DIMENSION = 3;
+	static const uint8 DIMENSION = 3;
 
-	static const int32 PRIM_SIZE = 1;
+	static const uint8 PRIM_SIZE = 1;
 
 	using MapTraits = MAP_TRAITS;
 	using MapType = MAP_TYPE;
